+++ src/main/java/cl/bicevida/MedioPago/domain/modelo/MedioPago_Modelo.java
@@ -24,9 +24,9 @@ import lombok.NoArgsConstructor;
 public class MedioPago_Modelo {
     @Id
     @Column(name = "id", nullable = false)
-    @Schema(required = true, implementation = Integer.class, example = "1")
+    @Schema(required = true, implementation = Long.class, example = "1")
     @JsonbProperty("id")
-    public Integer id;
+    public Long id;
 
     @ManyToOne
     @JoinColumn(name = "id_tipo_medio_pago")
