+++ src/main/java/cl/bicevida/MotivoReversa/domain/modelo/MotivoReversa_Modelo.java
@@ -18,9 +18,9 @@ import org.eclipse.microprofile.openapi.annotations.media.Schema;
 public class MotivoReversa_Modelo {
     @Id
     @Column(name = "id", nullable = false)
-    @Schema(required = true, implementation = Integer.class, example = "1")
+    @Schema(required = true, implementation = Long.class, example = "1")
     @JsonbProperty("id")
-    public Integer id;
+    public Long id;
 
     @Column(name = "descripcion")
     @Schema(required = true, implementation = String.class, example = "Motivo de reversa")
